requirements.txt
Django==2.0
django-localflavor==2.1
djangorestframework==3.10.3
djangorestframework-gis==0.14
psycopg2==2.8.4
python-stdnum==1.12
pytz==2019.3
six==1.11.0
sqlparse==0.3.0
validate-docbr==1.6.0
virtualenv==16.7.8
